Output status to stderr
This means you can use > to save the output if you want without having to specify an output file.
Also expect that the url to download will be the first non flag argument passed to the program.
This also requires that all flags are passed before the url to download.