Handle no scheme on URL
Assume that if no scheme is provided on the url to curly fry
then the user must have meant http. Also we reparse the URL
as without a scheme the Host field is not filled in. Throw
away the error on the second parse as it's taking as input
a String() on a url.URL, which must (I hope) be valid.