Add basic cookie jar reading and storing
This implementation uses the default stdlib cookie jar
which unfortunately doesn't provide full access to the
cookies in the cookiejar, so we are only able to save
cookies that are keyed for the URL in the request
equally any cookies set for the request are set for
the url of the request, not the url they might have
come from. The potential issue here would be when
following redirects, the answer will probably be
to implement our own cookiejar.