Headers, agent and data
This implements the headers option (multiple instances of) and tries
to copy the curl implementation as much as possible, although the
golang stdlib might now allow you unset some headers.
Also this support the -A option for setting the user agent string.
This should also handle all types of --data-x and encoding them
for sending to the server, to be honest I'm not 100% sure what the
difference is between data-binary and data-raw, should data binary
support the @filename for content? Essentially this looks to do what
we'd want.

resolves #11